Create cryptech.json
I really hope that your team will have the opportunity to cooperate with us. We are the only ones physically located in Ukraine, we have our own data center, and a huge experience in working with blockchain technologies, including the Cosmos SDK. Also, we supported the project from the first testnet.